+++ physicsnemo/datapipes/cae/domino_datapipe.py
@@ -107,7 +107,7 @@ class DoMINODataConfig:
         phase: Which phase of data to load ("train", "val", or "test").
         surface_variables: (Surface specific) Names of surface variables.
         surface_points_sample: (Surface specific) Number of surface points to sample per batch.
-        num__surface_neighbors: (Surface specific) Number of surface neighbors to consider for nearest neighbors approach.
+        num_surface_neighbors: (Surface specific) Number of surface neighbors to consider for nearest neighbors approach.
         resample_surfaces: (Surface specific) Whether to resample the surface before kdtree/knn. Not available if caching.
         resampling_points: (Surface specific) Number of points to resample the surface to.
         surface_sampling_algorithm: (Surface specific) Algorithm to use for surface sampling ("area_weighted" or "random").
@@ -658,16 +658,17 @@ class DoMINODataPipe(Dataset):
                 )
 
             # Fit the kNN (or KDTree, if CPU) on ALL points:
-            if self.array_provider == cp:
-                knn = cuml.neighbors.NearestNeighbors(
-                    n_neighbors=self.config.num_surface_neighbors,
-                    algorithm="rbc",
-                )
-                knn.fit(surface_coordinates)
-            else:
-                # Under the hood this is instantiating a KDTree.
-                # aka here knn is a type, not a class, technically.
-                interp_func = KDTree(surface_coordinates)
+            if self.config.num_surface_neighbors > 1:
+                if self.array_provider == cp:
+                    knn = cuml.neighbors.NearestNeighbors(
+                        n_neighbors=self.config.num_surface_neighbors,
+                        algorithm="rbc",
+                    )
+                    knn.fit(surface_coordinates)
+                else:
+                    # Under the hood this is instantiating a KDTree.
+                    # aka here knn is a type, not a class, technically.
+                    interp_func = KDTree(surface_coordinates)
 
             if self.config.sampling:
                 # Perform the down sampling:
@@ -700,22 +701,28 @@ class DoMINODataPipe(Dataset):
                 pos_normals_com_surface = pos_normals_com_surface[idx_surface]
 
                 # Now, perform the kNN on the sampled points:
-                if self.array_provider == cp:
-                    ii = knn.kneighbors(
-                        surface_coordinates_sampled, return_distance=False
-                    )
-                else:
-                    _, ii = interp_func.query(
-                        surface_coordinates_sampled, k=self.config.num_surface_neighbors
-                    )
+                if self.config.num_surface_neighbors > 1:
+                    if self.array_provider == cp:
+                        ii = knn.kneighbors(
+                            surface_coordinates_sampled, return_distance=False
+                        )
+                    else:
+                        _, ii = interp_func.query(
+                            surface_coordinates_sampled,
+                            k=self.config.num_surface_neighbors,
+                        )
 
-                # Pull out the neighbor elements.  Note that ii is the index into the original
-                # points - but only exists for the sampled points
-                # In other words, a point from `surface_coordinates_sampled` has neighbors
-                # from the full `surface_coordinates` array.
-                surface_neighbors = surface_coordinates[ii][:, 1:]
-                surface_neighbors_normals = surface_normals[ii][:, 1:]
-                surface_neighbors_sizes = surface_sizes[ii][:, 1:]
+                    # Pull out the neighbor elements.  Note that ii is the index into the original
+                    # points - but only exists for the sampled points
+                    # In other words, a point from `surface_coordinates_sampled` has neighbors
+                    # from the full `surface_coordinates` array.
+                    surface_neighbors = surface_coordinates[ii][:, 1:]
+                    surface_neighbors_normals = surface_normals[ii][:, 1:]
+                    surface_neighbors_sizes = surface_sizes[ii][:, 1:]
+                else:
+                    surface_neighbors = surface_coordinates
+                    surface_neighbors_normals = surface_normals
+                    surface_neighbors_sizes = surface_sizes
 
                 # We could index into these above the knn step too; they aren't dependent on that.
                 surface_normals = surface_normals[idx_surface]
@@ -1437,7 +1444,7 @@ def create_domino_dataset(
             model_type=cfg.model.model_type,
             bounding_box_dims=cfg.data.bounding_box,
             bounding_box_dims_surf=cfg.data.bounding_box_surface,
-            num_surface_neighbors=cfg.model.num_surface_neighbors,
+            num_surface_neighbors=cfg.model.num_neighbors_surface,
             resample_surfaces=cfg.model.resampling_surface_mesh.resample,
             resampling_points=cfg.model.resampling_surface_mesh.points,
             surface_sampling_algorithm=cfg.model.surface_sampling_algorithm,