+++ physicsnemo/utils/domino/utils.py
@@ -1140,8 +1140,12 @@ def area_weighted_shuffle_array(
     # Create index array for all available points
     point_indices = xp.arange(arr.shape[0])
 
-    selected_indices = xp.random.choice(
-        xp.asarray(point_indices), size=n_points, p=xp.asarray(sampling_probabilities)
+    if xp != np:
+        point_indices = point_indices.get()
+        sampling_probabilities = sampling_probabilities.get()
+
+    selected_indices = np.random.choice(
+        point_indices, n_points, p=sampling_probabilities
     )
     selected_indices = xp.asarray(selected_indices)